A2SG-39 refactor: separated routes by functions and cleanup db file
